+++ src/location/maps/tiled/qgeotiledmapdata.h
@@ -61,7 +61,7 @@ public:
 
     void setZoomLevel(qreal zoomLevel);
 
-    void pan(qreal dx, qreal dy);
+    void pan(int dx, int dy);
 
     void setCenter(const QGeoCoordinate &center);
     QGeoCoordinate center() const;
